Add flycheck::command
