+++ crates/rust-analyzer/src/flycheck.rs
@@ -4,10 +4,10 @@
 use std::{fmt, io, process::Command, time::Duration};
 
 use cargo_metadata::PackageId;
+use command::{FlycheckConfig, InvocationStrategy, Target};
 use crossbeam_channel::{select_biased, unbounded, Receiver, Sender};
 use ide_db::FxHashSet;
-use paths::{AbsPath, AbsPathBuf, Utf8PathBuf};
-use rustc_hash::FxHashMap;
+use paths::{AbsPath, AbsPathBuf};
 use serde::Deserialize as _;
 use serde_derive::Deserialize;
 
@@ -19,85 +19,7 @@ use triomphe::Arc;
 
 use crate::command::{CommandHandle, ParseFromLine};
 
-#[derive(Clone, Debug, Default, PartialEq, Eq)]
-pub(crate) enum InvocationStrategy {
-    Once,
-    #[default]
-    PerWorkspace,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(crate) struct CargoOptions {
-    pub(crate) target_tuples: Vec<String>,
-    pub(crate) all_targets: bool,
-    pub(crate) no_default_features: bool,
-    pub(crate) all_features: bool,
-    pub(crate) features: Vec<String>,
-    pub(crate) extra_args: Vec<String>,
-    pub(crate) extra_test_bin_args: Vec<String>,
-    pub(crate) extra_env: FxHashMap<String, String>,
-    pub(crate) target_dir: Option<Utf8PathBuf>,
-}
-
-#[derive(Clone, Debug)]
-pub(crate) enum Target {
-    Bin(String),
-    Example(String),
-    Benchmark(String),
-    Test(String),
-}
-
-impl CargoOptions {
-    pub(crate) fn apply_on_command(&self, cmd: &mut Command) {
-        for target in &self.target_tuples {
-            cmd.args(["--target", target.as_str()]);
-        }
-        if self.all_targets {
-            cmd.arg("--all-targets");
-        }
-        if self.all_features {
-            cmd.arg("--all-features");
-        } else {
-            if self.no_default_features {
-                cmd.arg("--no-default-features");
-            }
-            if !self.features.is_empty() {
-                cmd.arg("--features");
-                cmd.arg(self.features.join(" "));
-            }
-        }
-        if let Some(target_dir) = &self.target_dir {
-            cmd.arg("--target-dir").arg(target_dir);
-        }
-        cmd.envs(&self.extra_env);
-    }
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(crate) enum FlycheckConfig {
-    CargoCommand {
-        command: String,
-        options: CargoOptions,
-        ansi_color_output: bool,
-    },
-    CustomCommand {
-        command: String,
-        args: Vec<String>,
-        extra_env: FxHashMap<String, String>,
-        invocation_strategy: InvocationStrategy,
-    },
-}
-
-impl fmt::Display for FlycheckConfig {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            FlycheckConfig::CargoCommand { command, .. } => write!(f, "cargo {command}"),
-            FlycheckConfig::CustomCommand { command, args, .. } => {
-                write!(f, "{command} {}", args.join(" "))
-            }
-        }
-    }
-}
+pub(crate) mod command;
 
 /// Flycheck wraps the shared state and communication machinery used for
 /// running `cargo check` (or other compatible command) and providing
