+++ crates/rust-analyzer/src/reload.rs
@@ -32,7 +32,7 @@ use vfs::{AbsPath, AbsPathBuf, ChangeKind};
 
 use crate::{
     config::{Config, FilesWatcher, LinkedProject},
-    flycheck::{FlycheckConfig, FlycheckHandle},
+    flycheck::{command::FlycheckConfig, FlycheckHandle},
     global_state::{
         FetchBuildDataResponse, FetchWorkspaceRequest, FetchWorkspaceResponse, GlobalState,
     },
@@ -802,7 +802,7 @@ impl GlobalState {
         let sender = self.flycheck_sender.clone();
         let invocation_strategy = match config {
             FlycheckConfig::CargoCommand { .. } => {
-                crate::flycheck::InvocationStrategy::PerWorkspace
+                crate::flycheck::command::InvocationStrategy::PerWorkspace
             }
             FlycheckConfig::CustomCommand { ref invocation_strategy, .. } => {
                 invocation_strategy.clone()
@@ -810,7 +810,7 @@ impl GlobalState {
         };
 
         self.flycheck = match invocation_strategy {
-            crate::flycheck::InvocationStrategy::Once => {
+            crate::flycheck::command::InvocationStrategy::Once => {
                 vec![FlycheckHandle::spawn(
                     0,
                     sender,
@@ -820,7 +820,7 @@ impl GlobalState {
                     None,
                 )]
             }
-            crate::flycheck::InvocationStrategy::PerWorkspace => {
+            crate::flycheck::command::InvocationStrategy::PerWorkspace => {
                 self.workspaces
                     .iter()
                     .enumerate()
